Add: UIに表示するためのpresenterとComponentを追加

+++ Unlimited-Surviver/Assets/Project/Scripts/DOTS/Component/GameEntity/Health/HealthComponent.cs
@@ -4,8 +4,8 @@ namespace DOTS
 {
     public struct HealthComponent : IComponentData
     {
-        public uint MaxHealth;
+        public float MaxHealth;
 
-        public int Health;
+        public float Health;
     }
 }
